Added Additional DateTime extensions. Improved SetTime.
GetNext: give me next Friday after this date.
GetPrevious: give me previous friday before this date.
GetFirst: give me first Friday in this month, or first day of the month, or first Tuesday of the year.
GetLast: give me friday Friday in this month, or last day of the month, or first Tuesday of the year.
SetDate: change a specific portion of the date, for example, only the month, and keep the rest unchanged.