Fixed the issue of not detecting nullable reference support correctly.
Before this, It was assumed that as long as the VS is 2019, user can enable C# 8.0 language and the feature support was checked against VS version. But since V15.8 of visual studio, C# version became tied to framework. C# 8.0 is only supported in .net Standard 2.1 compatible frameworks and above, regardless of VS version. This means no C# 8.0 feature can be used in any current or future version of .net framework. The T4 template checks the target framework now to determine compatibility. Also, based on the framework, highest version of C# is always selected and cannot be changed using VS GUI.