14 Review IEnumerable Extension to add Array equivalent
AddRange and InsertRange were updated to have different foreach loop per type of the collection passed.

For ToHierarchy, the same approach was implemented, but it was not reasonable to copy the whole big content of the loop three times. A private method was created that contains all the body of the foreach loop, and compiler is instruct it to inline the helper method.